[apptools]Rename the test suite to apptools-android-tests
- The test suite can be build and run normally.
