+++ apptools/apptools-android-tests/COPYING
