+++ apptools/apptools-android-tests/README
@@ -1,25 +1,25 @@
 ## Introduction
 
-This test suite is for wrt-apptools-android-tests
+This test suite is for apptools-android-tests
 1. ./allpairs.py is create pkgName.py and report directory for package name
 
 ## Precondition
 
 1. Connect Android devices to your localhost
-2. Need to edit the file "wrt-apptools-android-tests/arch.txt" according to the type of test device.
+2. Need to edit the file "apptools-android-tests/arch.txt" according to the type of test device.
 If test device is "arm" platform, content of the file should be "arm", default value is "x86"
 3. The node.js, Android SDK, and git must be functional
-4. Download crosswalk-app-tools to wrt-apptools-android-tests/tools, then install npm
+4. Download crosswalk-app-tools to apptools-android-tests/tools, then install npm
   4.1 cd tools, then clone source code from https://github.com/crosswalk-project/crosswalk-app-tools
   4.2 Run commands: cd crosswalk-app-tools, then run: sudo npm install
   4.3 Download Crosswalk from https://origin-download.01.org/crosswalk/releases/crosswalk/android/stable/ to same path with crosswalk-app-tools
 
 ## Test Steps
 
-1. unzip wrt-apptools-android-tests<version>.zip -d [testprefix-path]
-2. cd [testprefix-path]/opt/wrt-apptools-android-tests/
+1. unzip apptools-android-tests<version>.zip -d [testprefix-path]
+2. cd [testprefix-path]/opt/apptools-android-tests/
 3. run test case
-   testkit-lite -f [testprefix-path]/opt/wrt-apptools-android-tests/tests.xml -A -o [testprefix-path]/opt/wrt-apptools-android-tests/result.xml --comm localhost --testenvs "DEVICE_ID=Medfield3C6DFF2E;CONNECT_TYPE=adb" --testprefix=[testprefix-path], DEVICE_ID can also be multiple ids like "DEVICE_ID=Medfield3C6DFF2E,Medfield3C6DFF00".
+   testkit-lite -f [testprefix-path]/opt/apptools-android-tests/tests.xml -A -o [testprefix-path]/opt/apptools-android-tests/result.xml --comm localhost --testenvs "DEVICE_ID=Medfield3C6DFF2E;CONNECT_TYPE=adb" --testprefix=[testprefix-path], DEVICE_ID can also be multiple ids like "DEVICE_ID=Medfield3C6DFF2E,Medfield3C6DFF00".
 Query device id by command "adb devices -l" in host.
 
 ## Authors:
