+++ apptools/apptools-android-tests/arch.txt
