+++ apptools/apptools-android-tests/report/cmd.txt
@@ -1,0 +1,16 @@
+positive1	crosswalk-app create org.xwalk.tests --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive2	crosswalk-app create org.xwalk.t1234 --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive3	crosswalk-app create org.example._xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive4	crosswalk-app create org.example.xwal_ --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive5	crosswalk-app create org.example.te_st --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive6	crosswalk-app create org.xwalk.Tests --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive7	crosswalk-app create or_g.example.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive8	crosswalk-app create org000.example.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive9	crosswalk-app create _org.example.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+positive10	crosswalk-app create org.example123.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative11	crosswalk-app create org.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative12	crosswalk-app create test --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative13	crosswalk-app create org.example.1234test --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative14	crosswalk-app create org.example.1234 --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative15	crosswalk-app create 123org.example.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
+negative16	crosswalk-app create org.123example.xwalk --crosswalk=/opt/apptools-android-tests/tools/crosswalk*.zip
