+++ apptools/apptools-android-tests/suite.json
@@ -25,5 +25,5 @@
             "blacklist": []
         }
     },
-    "pkg-name": "wrt-apptools-android-tests"
+    "pkg-name": "apptools-android-tests"
 }
