+++ apptools/apptools-android-tests/tools/README
